Adds Hamilton to orchestrators
Hamilton is a lightweight micro-orchestration system that provides a "dbt" for python story; your code is always unit testable, integration testing friendly, and comes with lineage out-of-the-box.

Hamilton does not replace things like Airflow or Dagster because it is just a library. It runs anywhere python runs. 

Because Hamilton is lightweight you can use it for data processing, machine learning, and LLM workflows. These "dataflows" are then portable and can be run in a notebook, python script, Airflow python task, pyspark job, a python web-service, etc.